### 2025-07-08 #8555 @editorialbot
[PRE REVIEW]: Spectralmatch: relative radiometric normalization toolkit for raster mosaics and time series
**Submitting author:** @cankanoa (<a href="http://orcid.org/0009-0009-5520-1911">Kanoa Lindiwe</a>)
**Repository:** https://github.com/spectralmatch/spectralmatch
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
