### 2025-09-05 #8555 @Digdgeo
Hi! I'd be happy to review this submission for JOSS. I have experience with Remote Sensing and am familiar with the review criteria.

Please let me know the expected timeline and any specific aspects you'd like me to focus on during the review.

Thanks for considering me as a reviewer.
