### 2025-09-12 #8974 @editorialbot
[REVIEW]: Spectralmatch: relative radiometric normalization toolkit for raster mosaics and time series
**Submitting author:** @cankanoa (<a href="http://orcid.org/0009-0009-5520-1911">Kanoa Lindiwe</a>)
**Repository:** https://github.com/spectralmatch/spectralmatch
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.4

## Notes

- @simonrp84 is not expected to complete review until near the end of the 6 weeks
