### 2025-09-25 #8974 @simonrp84
I've now read through the paper and overall it's good, but I have a few minor points:
1) The paper isn't clear about exactly what data the algorithm can be applied to. Should it be applied only to images from one platform (i.e: WV3) or can it be applied equally well across multiple platforms (i.e: WV3 + Sentinel-2)?
2) The statement of need section discusses things like atmospheric and geometric effects but doesn't mention that these can be, more accurately, adjusted for with atmospheric or BRDF correction. A brief mention of these would be appropriate.
3) The statement of need mentions both atmospheric and adjacency effects, these are the same thing - you can remove adjacency from this list.
4) Parallelisation is mentioned, if possible it'd be nice to see an indication of potential speed-ups or efficiency gains enabled by this parallelisation.
