### 2025-10-06 2a320e8
Address JOSS reviewers by adding versioning, min overlap size, and improving joss paper
Imported versioning from pyproject.toml to manage package metadata. Introduced a `min_overlap_size` parameter in global regression for enhanced control over matching thresholds. Updated documentation and added a new reference to improve clarity and provide additional context for implemented methods. Improved paper.md for JOSS reviewers.
