+++ research/joss/paper.bib
@@ -50,6 +50,16 @@
   year = {2024}
 }
 
+@article{Shen:2025,
+  title = {BRDF Correction in Landsat Directional Reflectance Data by Combining Optimal BRDF Archetype With Spectral Characteristics},
+  author = {A. Shen and J. Shao and X. Qi and Z. Zhu and Y. Dai and Q. Zeng},
+  journal = {IEEE Journal of Selected Topics in Applied Earth Observations and Remote Sensing},
+  year = {2025},
+  volume = {18},
+  pages = {17592--17609},
+  doi = {10.1109/JSTARS.2025.3586463},
+}
+
 @article{Theiler:2019,
   doi = {10.1109/MGRS.2019.2890997},
   author = {Theiler, J. and Ziemann, A. and Matteoli, S. and Diani, M.},
