### 2025-10-08 e2f98bd
Improve JOSS paper for clarity and detail
Updated the text to enhance clarity and readability, including improved descriptions of algorithms, methods, and functionalities. Added specific details about computational optimizations and dataset handling for large-scale processing.
