### 2025-10-09 ef36540
Fix typos and improve clarity in JOSS paper
Corrected several typographical errors and refined phrasing in the JOSS paper for better readability and accuracy. Adjustments include fixes to spelling inconsistencies, formatting, and clearer explanations of functionalities and methods.
