+++ research/joss/paper.md
@@ -30,25 +30,25 @@ affiliations:
 date: 8 May 2025
 bibliography: paper.bib
 ---
-# Summary  
-Spectralmatch provides algorithms to perform relative radiometric normalization (RRN) to enhance spectral consistency across raster mosaics and time series. It is built for geoscientific use, with a sensor- and unit-agnostic design, optimized for automation and efficiency on arbitrarily many images and bands, and works well with Very High Resolution Imagery (VHRI) as it does not require pixel co-registration. Its current matching algorithms are inspired by @Yu:2017, which include global regression and local block adjustment that minimize inter-image variability without relying on ancillary data. The impact of these functions on spectral consistency is illustrated in \autoref{fig:1}. The software supports cloud and vegetation masking, pseudo invariant feature (PIF) based exclusion, seamline network generation, raster merging, and plotting statistics. The toolkit is available as an open-source Python library, command line interface, and QGIS plugin.
+# Summary
+Spectralmatch provides algorithms to perform relative radiometric normalization (RRN) to enhance spectral consistency across raster mosaics and time series. It is built for geoscientific use, with a sensor- and unit-agnostic design, optimized for automation and efficiency on arbitrarily many images and bands, and works well with Very High Resolution Imagery (VHRI) as it does not require pixel co-registration. Its current matching algorithms are inspired by @Yu:2017, which include global regression and local block adjustment which minimize inter-image variability without relying on ancillary data. The impact of these functions on spectral consistency is illustrated in \autoref{fig:1}. The software supports cloud and vegetation masking, pseudo invariant feature (PIF) based exclusion, seamline network generation, raster merging, and plotting statistics. The toolkit is available as an open-source Python library, command line interface, and QGIS plugin.
   
-# Statement of Need  
-Remote sensing relies on mosaics to broaden spatial coverage by merging images across space, and on time series to extend temporal coverage by linking images over time. However, both are affected by inter-image spectral variability, caused by differences in the atmosphere, illumination, surface condition, acquisition geometry, and other complications [@Theiler:2019]. These factors introduce inconsistencies, reduce accuracy in image analysis, and complicate the detection of actual environmental changes. To address these issues, researchers have explored two main correction approaches in the literature: absolute radiometric correction and RRN, or a combination of both [@Hu:2011]. The absolute approach corrects for the aforementioned inaccuracies with algorithms involving atmospheric correction and bidirectional reflectance distribution functions [@Shen:2025] and gains accuracy from in-situ measurements, which may not exist or be difficult to obtain for specific images [@Canty:2004]. Conversely, the relative approach applies algorithms to minimize apparent spectral differences between images, matching them for consistent analysis, rather than determining true spectral values or relying on ancillary data.
+# Statement of Need
[…]
