### 2025-10-09 ac7c411
Update paper.md and image for paper.
Refined the explanation of the local block adjustment algorithm for greater clarity and detail. Highlighted the enhanced scalability benefits, particularly for larger image datasets, and specified improvements in multiprocessing speed gains. Updated the image for the paper to be only the overlap area.
