+++ research/joss/paper.md
@@ -41,12 +41,12 @@ Researchers have examined various algorithms for performing RRN [@Vorovencii:201
 # Implemented RRN Methods
 The current matching algorithm uses a two-step approach involving global regression and local block adjustment following the methods of @Yu:2017. The approach is suitable to match imagery from the same sensor or from multiple sensors with comparable wavelengths and resolution—for example, between satellites (Sentinel-PlanetScope) or between drones (Zenmuse P1-Mavic 3 Multispectral). The global regression algorithm adjusts brightness and contrast across overlapping images to reduce spectral differences. It first detects overlapping image pairs and computes per-band statistics (mean and standard deviation) within those regions. Using these statistics, a least-squares regression system is constructed to solve for per-image, per-band scale and offset parameters that minimize radiometric differences in overlapping areas. This approach aims to minimize brightness and contrast differences across images while preserving global consistency and aligning the spectral profiles of images to a central tendency, specific image, or set of images via custom-weighted mean and standard deviation constraints.
   
-The local block adjustment algorithm applies block-wise radiometric correction to individual satellite images based on local differences from a reference mosaic. The method divides the combined extent of all input images into spatial blocks and calculates local mean statistics for each block. Each image is then locally adjusted using interpolated adaptive gamma normalization to align with the global reference mosaic. This allows radiometric consistency across spatially heterogeneous scenes on a block scale. Both global and local algorithms support nodata-aware processing for images of irregular shapes and internal gaps and vector PIF masking applied at the matching solution stage. For large datasets, the library supports cloud optimized geotiff outputs, saving and loading of intermediate steps, and efficient windowing and parallelization. For example, on a computer with 16GB of memory and 10 M1 cores, multiprocessing increased processing speed by up to 43%, depending on the image size and computer hardware.
+The local block adjustment algorithm applies block-wise radiometric correction to individual satellite images based on local differences from a reference mosaic. The method divides the combined extent of all input images into spatial blocks and calculates local mean statistics for each block. Each image is then locally adjusted using interpolated adaptive gamma normalization to align with the global reference mosaic. This allows radiometric consistency across spatially heterogeneous scenes on a block scale. Both global and local algorithms support nodata-aware processing for images of irregular shapes and internal gaps and vector PIF masking applied at the matching solution stage. For large datasets, the library supports cloud optimized geotiff outputs, saving and loading of intermediate steps, and efficient windowing and parallelization. For example, on a computer with 16GB of memory and 10 M1 cores, multiprocessing increased processing speed by up to 43%, with larger images showing larger increases.
 
 Various helper functions support the creation of cloud masks, non-vegetation PIFs, generating seamline networks, merging images, and basic figures. Cloud masking utilities enable the generation of binary masks using [OmniCloudMask](https://github.com/DPIRD-DMA/OmniCloudMask) by @Wright:2025, followed by post-processing and vectorization. Vegetation masking utilities use NDVI-based thresholds, followed by post-processing and vectorization. The created masks can be used to mask input images or withhold pixels from analysis. Seamline generation utilities use Voronoi-based centerlines, following the methodology of @Yuan:2023. Statistical utilities can generate basic figures comparing image spectral profiles before and after matching to evaluate radiometric changes. Raster merging utilities combine the final images into a seamless mosaic.
 
 # Figures
[…]
