### 2025-10-20 spectralmatch/spectralmatch#49 @simonrp84
Cannot install on windows
When I try to install `spectralmatch` on windows I get this error:
```
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [22 lines of output]
      + C:\Users\simon.proud\AppData\Local\miniforge3\envs\testspec\python.exe C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d\vendored-meson\meson\meson.py setup C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d\.mesonpy-7pycd4zw -Dbuildtype=release -Db_ndebug=if-release -Db_vscrt=md --native-file=C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d\.mesonpy-7pycd4zw\meson-python-native-file.ini
      The Meson build system
      Version: 1.5.2
      Source dir: C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d
      Build dir: C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d\.mesonpy-7pycd4zw
      Build type: native build
      WARNING: failed to process netrc file: bad follower token "'password" (C:\Users\simon.proud\.netrc, line 3).
      Project name: NumPy
      Project version: 2.2.3
      WARNING: Failed to activate VS environment: Could not find C:\Program Files (x86)\Microsoft Visual Studio\Installer\vswhere.exe

      ..\meson.build:1:0: ERROR: Unknown compiler(s): [['icl'], ['cl'], ['cc'], ['gcc'], ['clang'], ['clang-cl'], ['pgcc']]
      The following exception(s) were encountered:
      Running `icl ""` gave "[WinError 2] The system cannot find the file specified"
      Running `cl /?` gave "[WinError 2] The system cannot find the file specified"
      Running `cc --version` gave "[WinError 2] The system cannot find the file specified"
      Running `gcc --version` gave "[WinError 2] The system cannot find the file specified"
      Running `clang --version` gave "[WinError 2] The system cannot find the file specified"
      Running `clang-cl /?` gave "[WinError 2] The system cannot find the file specified"
      Running `pgcc --version` gave "[WinError 2] The system cannot find the file specified"

      A full log can be found at C:\Users\simon.proud\AppData\Local\Temp\pip-install-022kffln\numpy_c5e7351db756423cb4e5b9be6b90260d\.mesonpy-7pycd4zw\meson-logs\meson-log.txt
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```

This is because of the missing VC++ build tools. Unfortunately, on my laptop - as with many other corporate systems - there are strict security controls and it is not possible to install the build tools.

Is there an alternative solution for building on Windows?
