### 2025-10-22 #8974 @cankanoa
Aloha all,

@simonrp84 has created three GitHub [issues](https://github.com/spectralmatch/spectralmatch/issues) two days ago. I will address these shortly. I'm not sure if there is other comments as part of their review.

Thanks for the issues @simonrp84
