### 2025-10-23 8d1bdd9
Add size-based weighting option for global regression
Introduced `weight_by_size` parameter to allow weighting constraints by overlap size during global regression calculations. This change provides flexibility for users to enable or disable size-based weighting, enhancing the control over constraint balancing. Also simplified debug logging and loop structures for readability.
