### 2025-10-23 fdd28e2
Update Python and GDAL version constraints in docs
Clarified that the plugin supports Python versions between 3.10 and 3.12. Adjusted GDAL requirement to "≥ 3.10.2" for consistency. Updated installation instructions accordingly.
