+++ docs/installation.md
@@ -5,7 +5,7 @@
 ## Installation as a QGIS Plugin
 ### 1. Get QGIS
 [Download](https://qgis.org/download/) and install QGIS.
-> This plugin requires python>=3.10. QGIS ships with different versions of Python, to check, in the QGIS menu, go to QGIS>About gis. If your version of Python is not supported, you can update your QGIS (if available) or install it containerized with conda: `conda create --name qgis_env python=3.12`, `conda activate qgis_env`, `conda install -c conda-forge qgis`, then `qgis` to start the program.
+> This plugin requires Python ≥ 3.10 and ≤ 3.12. QGIS ships with different versions of Python, to check, in the QGIS menu, go to QGIS>About gis. If your version of Python is not supported, you can update your QGIS (if available) or install it containerized with conda: `conda create --name qgis_env python=3.12`, `conda activate qgis_env`, `conda install -c conda-forge qgis`, then `qgis` to start the program.
 
 ### 2. Install spactalmatch QGIS plugin
 - Go to Plugins → Manage and Install Plugins…
@@ -22,14 +22,14 @@ The plugin will attempt to automatically install all Python dependencies that it
 ### 1. System requirements
 Before installing, ensure you have the following system-level prerequisites:
 
-- Python ≥ 3.10
+- Python ≥ 3.10 and ≤ 3.12
 - PROJ ≥ 9.3
-- GDAL = 3.10.2
+- GDAL ≥ 3.10.2
 - pip
 
 An easy way to install these dependancies is to use [Miniconda](https://www.anaconda.com/docs/getting-started/miniconda/install#quickstart-install-instructions):
 ```bash
-conda create -n spectralmatch python=3.10 "gdal=3.10.2" "proj>=9.3" -c conda-forge
+conda create -n spectralmatch python=3.12 "gdal>=3.10.2" "proj>=9.3" -c conda-forge
 conda activate spectralmatch
 ```
 
@@ -56,13 +56,13 @@ cd spectralmatch
 ### 2. System requirements
 Before installing, ensure you have the following system-level prerequisites:
 
-- Python ≥ 3.10
+- Python ≥ 3.10 and ≤ 3.12
 - PROJ ≥ 9.3
 - GDAL = 3.10.2
 
 An easy way to install these dependancies is to use [Miniconda](https://www.anaconda.com/docs/getting-started/miniconda/install#quickstart-install-instructions):
 ```bash
-conda create -n spectralmatch python=3.10 "gdal=3.10.2" "proj>=9.3" -c conda-forge
+conda create -n spectralmatch python=3.12 "gdal>=3.10.2" "proj>=9.3" -c conda-forge
 conda activate spectralmatch
 ```
 
