### 2025-10-23 bcf37e2
Restrict Python version compatibility in README.
Clarified that the library supports Python versions ≥ 3.10 and ≤ 3.12. This ensures accurate prerequisites for users during installation.
