### 2025-11-03 #8974 @ethanwhite
Thanks for opening those issues @simonrp84 and for addressing them so quickly @cankanoa! It looks like we're probably currently hung up on the installation issue that @simonrp84 was having so I've added a little follow up in https://github.com/spectralmatch/spectralmatch/issues/49 to try to help move that forward.
