### 2025-11-08 a72b3e9
Add Windows test workflow and support multiple Python versions
This update extends the CI workflows to include a new `test-windows` job, adding testing support on Windows with Python versions 3.10, 3.11, and 3.12. Minor fixes were also applied to ensure consistency in syntax and configuration flags (e.g., correcting boolean values).
