+++ .github/workflows/main.yml
@@ -4,12 +4,13 @@ on:
   push:
     branches:
       - main
+      - test
   pull_request:
     branches:
       - main
 
 jobs:
-  test:
+  test-linux:
     runs-on: ubuntu-latest
 
     strategy:
@@ -26,7 +27,7 @@ jobs:
           auto-update-conda: true
           python-version: ${{ matrix.python-version }}
           channels: conda-forge
-          conda-remove-defaults: True
+          conda-remove-defaults: true
 
       - name: Create Conda environment with Python and GDAL
         run: |
@@ -36,7 +37,7 @@ jobs:
         shell: bash -l {0}
         run: |
           conda activate spectralmatch_env
-          pip install -e ".[dev]" 
+          pip install -e ".[dev]"
 
       - name: Run tests
         shell: bash -l {0}
@@ -48,3 +49,40 @@ jobs:
         uses: codecov/codecov-action@v5
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
+
+  test-windows:
+    runs-on: windows-latest
+
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+
+    defaults:
+      run:
+        shell: bash -l {0}
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Install Miniconda
+        uses: conda-incubator/setup-miniconda@v3
+        with:
+          auto-update-conda: true
+          python-version: ${{ matrix.python-version }}
+          channels: conda-forge
+          conda-remove-defaults: true
+
+      - name: Create Conda environment with GDAL (Windows)
+        run: |
+          conda create --name spectralmatch_env python=${{ matrix.python-version }} "gdal=3.10.2" "proj>=9.3" --override-channels -c conda-forge -v
+
+      - name: Activate Conda environment and install package
+        run: |
+          conda activate spectralmatch_env
+          pip install -e ".[dev]"
+
+      - name: Run tests
+        run: |
+          conda activate spectralmatch_env
+          pytest
