### 2025-11-08 34697df
Update installation instructions and add pixi installation guide
Improved installation guidance in README and documentation for clarity, including updated QGIS plugin steps and pip method prerequisites. Added a new section in the docs for installation via pixi, providing an alternative for handling dependencies seamlessly.
