### 2025-11-08 spectralmatch/spectralmatch#49 @cankanoa
Aloha @ethanwhite,

Thank you for your help.

I've added windows install tests to the main branch in this commit: https://github.com/spectralmatch/spectralmatch/pull/53. They all pass.

I've added the conda install commands to the README as I agree that including them in here makes it easier for most users. I've added the pixi install instructions to the install.md file. Both of these changes are merged into main here: https://github.com/spectralmatch/spectralmatch/pull/54.

Thanks  @henrykironde and @alexblochel for helping with this.
