### 2025-11-10 spectralmatch/spectralmatch#49 @simonrp84
I'm still attempting to find a solution to this, it installs on linux but not on windows and I'm reluctant to continue with the review unless we can get it working on windows. This week I'm overwhelmed with work but have asked some colleagues to help test potential solutions for Spectralmatch, so hopefully they will come up with something. If not, I'll have lots more time next week to test things out.
