### 2025-11-11 spectralmatch/spectralmatch#49 @cankanoa
Thanks @simonrp84 for your help with this. I agree that doing what @ethanwhite recommended will help debug the issue and hopefully figure out the root cause.
