+++ spectralmatch/match/global_regression.py
@@ -1,28 +1,21 @@
-import rasterio
-import fiona
 import os
 import numpy as np
 import sys
 import json
-from skimage.transform import resize
+import tempfile
+import re
 
+from osgeo import gdal
+gdal.UseExceptions()
 from concurrent.futures import as_completed
-from typing import List, Tuple, Literal
+from typing import List, Dict
 from numpy import ndarray
 from scipy.optimize import least_squares
-from rasterio.windows import Window
-from rasterio.transform import rowcol
-from rasterio.features import geometry_mask
-from rasterio.coords import BoundingBox
-from tqdm import tqdm
 
-from ..handlers import _resolve_paths, _get_nodata_value, _check_raster_requirements
-from ..utils_multiprocessing import (
-    _resolve_parallel_config,
-    _resolve_windows,
-    _get_executor,
-    WorkerContext,
-)
+from ..utils import create_masked_vrts, _set_gdal_cache, _set_gdal_workers, _resolve_gdal_dtype, _resolve_window_size, \
+    _get_gdal_bounds, _get_valid_count
+from ..handlers import _resolve_paths, _resolve_nodata_value, _check_raster_requirements
+from ..utils_multiprocessing import (_resolve_parallel_config, _get_executor)
 from ..types_and_validation import Universal, Match
 
 
@@ -35,8 +28,11 @@ def global_regression(
     vector_mask: Universal.VectorMask = None,
     debug_logs: Universal.DebugLogs = False,
     custom_nodata_value: Universal.CustomNodataValue = None,
-    image_parallel_workers: Universal.ImageParallelWorkers = None,
-    window_parallel_workers: Universal.WindowParallelWorkers = None,
+    cache: Universal.Cache = None,
+    image_threads: Universal.Threads = None,
+    io_threads: Universal.Threads = None,
+    tile_threads: Universal.Threads = None,
+    estimate_stats: bool = True,
     window_size: Universal.WindowSize = None,
     save_as_cog: Universal.SaveAsCog = False,
     specify_model_images: Match.SpecifyModelImages = None,
@@ -44,7 +40,6 @@ def global_regression(
     custom_std_factor: float = 1.0,
     save_adjustments: str | None = None,
     load_adjustments: str | None = None,
-    min_overlap_size: int | None = None,
 ) -> list:
     """
     Performs global radiometric normalization across overlapping images using least squares regression.
@@ -52,21 +47,23 @@ def global_regression(
     Args:
         input_images (str | List[str], required): Defines input files from a glob path, folder, or list of paths. Specify like: "/input/files/*.tif", "/input/folder" (assumes *.tif), ["/input/one.tif", "/input/two.tif"].
         output_images (str | List[str], required): Defines output files from a template path, folder, or list of paths (with the same length as the input). Specify like: "/input/files/$.tif", "/input/folder" (assumes $_Global.tif), ["/input/one.tif", "/input/two.tif"].
-        calculation_dtype (str, optional): Data type used for internal calculations. Defaults to "float32".
+        calculation_dtype (str, optional): Precision for internal calculations. Defaults to "float32".
         output_dtype (str | None, optional): Data type for output rasters. Defaults to input image dtype.
         vector_mask (Tuple[Literal["include", "exclude"], str, Optional[str]] | None): Mask to limit stats calculation to specific areas in the format of a tuple with two or three items: literal "include" or "exclude" the mask area, str path to the vector file, optional str of field name in vector file that *includes* (can be substring) input image name to filter geometry by. Loaded stats won't have this applied to them. The matching solution is still applied to these areas in the output. Defaults to None for no mask.
-        debug_logs (bool, optional): If True, prints debug information and constraint matrices. Defaults to False.
[…]
