### 2025-11-15 b465753
Refactor global regression to optimize GDAL functionality
Refactored `global_regression.py` to use GDAL VRTs, removing Rasterio in favor of more efficient GDAL workflows. Introduced modular handling for GDAL threads, datasets, and caching, while simplifying multiprocessing configurations for image and tile-level processing. Added support for advanced options like Cloud-Optimized GeoTIFFs and optimized VRT-based adjustments. Removed obsolete code paths and improved readability.
