+++ spectralmatch/match/local_block_adjustment.py
@@ -1,28 +1,17 @@
 import math
-import os
-import warnings
-
+import tempfile
 import numpy as np
-import rasterio
-import traceback
-import fiona
+import os
 
-from scipy.ndimage import map_coordinates, gaussian_filter
-from rasterio.windows import Window
-from rasterio.transform import from_origin
-from rasterio.features import geometry_mask
+from scipy.ndimage import gaussian_filter
 from concurrent.futures import as_completed
 from typing import Tuple, Optional, List, Literal
-from multiprocessing import shared_memory
+from osgeo import gdal, osr
 
-from ..handlers import _resolve_paths, _get_nodata_value, _check_raster_requirements
-from ..utils_multiprocessing import (
-    _resolve_parallel_config,
-    _resolve_windows,
-    _get_executor,
-    WorkerContext,
-    file_lock,
-)
+from ..utils import create_masked_vrts, _set_gdal_cache, _set_gdal_workers, _resolve_gdal_dtype, _resolve_window_size, \
+    _gdal_dtype_str_to_enum, _get_valid_count
+from ..handlers import _resolve_paths, _resolve_nodata_value, _check_raster_requirements
+from ..utils_multiprocessing import _resolve_parallel_config, _get_executor
 from ..types_and_validation import Universal, Match
 
 
@@ -35,13 +24,15 @@ def local_block_adjustment(
     vector_mask: Universal.VectorMask = None,
     debug_logs: Universal.DebugLogs = False,
     custom_nodata_value: Universal.CustomNodataValue = None,
-    image_parallel_workers: Universal.ImageParallelWorkers = None,
-    window_parallel_workers: Universal.WindowParallelWorkers = None,
-    window_size: Universal.WindowSizeWithBlock = None,
+    cache: Universal.Cache = None,
+    image_threads: Universal.Threads = None,
+    io_threads: Universal.Threads = None,
+    tile_threads: Universal.Threads = None,
+    window_size: Universal.WindowSize = None,
     save_as_cog: Universal.SaveAsCog = False,
     number_of_blocks: int | Tuple[int, int] | Literal["coefficient_of_variation"] = 100,
     alpha: float = 1.0,
-    correction_method: Literal["gamma", "linear"] = "gamma",
+    correction_method: Literal["gamma", "linear"] = "linear",
     save_block_maps: Tuple[str, str] | None = None,
     load_block_maps: (
         Tuple[str, List[str]] | Tuple[str, None] | Tuple[None, List[str]] | None
@@ -54,16 +45,18 @@ def local_block_adjustment(
 
     Args:
         input_images (str | List[str], required): Defines input files from a glob path, folder, or list of paths. Specify like: "/input/files/*.tif", "/input/folder" (assumes *.tif), ["/input/one.tif", "/input/two.tif"].
-        output_images (str | List[str], required): Defines output files from a template path, folder, or list of paths (with the same length as the input). Specify like: "/input/files/$.tif", "/input/folder" (assumes $_Local.tif), ["/input/one.tif", "/input/two.tif"].
+        output_images (str | List[str], required): Defines output files from a template path, folder, or list of paths (with the same length as the input). Specify like: "/input/files/$.tif", "/input/folder" (assumes $_Global.tif), ["/input/one.tif", "/input/two.tif"].
         calculation_dtype (str, optional): Precision for internal calculations. Defaults to "float32".
         output_dtype (str | None, optional): Data type for output rasters. Defaults to input image dtype.
[…]
