### 2025-11-15 f778ede
Refactor seamline generation and add AOI clipping support
Streamlined seamline generation by modularizing polygon handling and replacing rasterio with GDAL for efficiency. Added support for AOI-based clipping of seamlines and improved Voronoi seed point generation. Includes extensive code cleanup and enhanced geo-processing utilities.
