+++ spectralmatch/seamline/voronoi_center_seamline.py
@@ -1,26 +1,28 @@
+import math
 import os
-import rasterio
-import numpy as np
 import networkx as nx
 import fiona
 
 from fiona.errors import DriverError
-from rasterio.features import shapes
-from affine import Affine
-from shapely.geometry import MultiPoint
+from shapely.affinity import affine_transform
+from shapely.wkt import loads as wkt_loads
 from shapely.geometry import (
-    shape,
     Polygon,
     LineString,
     mapping,
     GeometryCollection,
     Point,
+    MultiPoint,
+    shape
 )
-from shapely.ops import split, voronoi_diagram
+from fiona import open as fopen
+from shapely.ops import split, voronoi_diagram, unary_union
 from itertools import combinations
-from typing import List, Tuple
+from typing import List
+from osgeo import gdal, ogr, osr
+gdal.UseExceptions()
 
-from ..handlers import search_paths, _resolve_paths
+from ..handlers import _resolve_paths
 from ..types_and_validation import Universal
 
 
@@ -28,6 +30,7 @@ def voronoi_center_seamline(
     input_images: Universal.CreateInFolderOrListFiles,
     output_mask: str,
     *,
+    aoi_path: str | None = None,
     image_field_name: str = "image",
     min_point_spacing: float = 10,
     min_cut_length: float = 0,
@@ -40,6 +43,7 @@ def voronoi_center_seamline(
     Args:
         input_images (str | List[str], required): Defines input files from a glob path, folder, or list of paths. Specify like: "/input/files/*.tif", "/input/folder" (assumes *.tif), ["/input/one.tif", "/input/two.tif"].
         output_mask (str): Output path for the final seamline polygon vector file.
+        aoi_path (str, optional): Path to an AOI vector file to clip overlapping image polygons; default is None.
         min_point_spacing (float, optional): Minimum spacing between Voronoi seed points; default is 10.
         min_cut_length (float, optional): Minimum cutline segment length to retain; default is 0.
         debug_logs (Universal.DebugLogs, optional): Enables debug print statements if True; default is False.
@@ -61,17 +65,27 @@ def voronoi_center_seamline(
 
     emps = []
     crs = None
-    for p in input_image_paths:
-        mask, transform = _read_mask(p, debug_logs)
-        emp = _seamline_mask(mask, transform, debug_logs)
+    for path in input_image_paths:
+        emp = _emp_polygon_from_image(path)
         emps.append(emp)
         if crs is None:
-            with rasterio.open(p) as src:
-                crs = src.crs
+            ds = gdal.Open(path, gdal.GA_ReadOnly)
+            crs = ds.GetProjectionRef()
+            ds = None
 
     for i, emp in enumerate(emps):
         if debug_logs:
             print(f"EMP{i}: area={emp.area:.2f}, bounds={emp.bounds}")
+
+    if os.path.exists(debug_vectors_path): os.remove(debug_vectors_path)
+    if debug_vectors_path:
+        _save_emp_outlines(
+            emps,
+            input_image_paths,
+            debug_vectors_path,
+            crs,
+            image_field_name=image_field_name,
+        )
 
     cuts: List[LineString] = []
     for i, (a, b) in enumerate(combinations(emps, 2)):
@@ -99,7 +113,7 @@ def voronoi_center_seamline(
             debug_vectors_path,
             "w",
             driver="GPKG",
-            crs=crs,
+            crs_wkt=crs,
             schema=schema,
             layer="cutlines",
         ) as dst:
[…]
