### 2025-11-15 a22265d
Refactor raster processing to use GDAL and enhance threshold logic
Replaced Rasterio with GDAL for raster data processing to improve performance and flexibility. Streamlined cloud mask generation, NDVI calculations, and added support for muparser expressions. Improved multiprocessing control and parameterized window size for enhanced memory efficiency.
