+++ spectralmatch/mask/mask.py
@@ -1,23 +1,18 @@
+import tempfile
 import re
-import rasterio
 import os
 import numpy as np
 
+from typing import List
 from omnicloudmask import predict_from_array
-from rasterio.enums import Resampling
-from rasterio.transform import from_origin
 from concurrent.futures import as_completed
-from typing import Any
-from typing import Tuple
+from osgeo import gdal
+gdal.UseExceptions()
 
 from ..types_and_validation import Universal
-from ..handlers import _resolve_paths, _resolve_output_dtype, _resolve_nodata_value
-from ..utils_multiprocessing import (
-    _resolve_parallel_config,
-    _get_executor,
-    WorkerContext,
-    _resolve_windows,
-)
+from ..handlers import _resolve_paths, _resolve_nodata_value, _check_raster_requirements
+from ..utils_multiprocessing import _resolve_parallel_config, _get_executor
+from ..utils import _set_gdal_cache, _set_gdal_workers, _resolve_gdal_dtype, _resolve_window_size
 
 
 def create_cloud_mask_with_omnicloudmask(
@@ -29,7 +24,7 @@ def create_cloud_mask_with_omnicloudmask(
     *,
     down_sample_m: float = None,
     debug_logs: Universal.DebugLogs = False,
-    image_parallel_workers: Universal.ImageParallelWorkers = None,
+    image_threads: Universal.Threads = None,
     omnicloud_kwargs: dict | None = None,
 ):
     """
@@ -41,9 +36,9 @@ def create_cloud_mask_with_omnicloudmask(
         red_band_index (int): Index of red band in the image.
         green_band_index (int): Index of green band in the image.
         nir_band_index (int): Index of NIR band in the image.
-        down_sample_m (float, optional): If set, resamples input to this resolution in meters.
+        down_sample_m (float, optional): If set, resamples input to this resolution in meters. Recommended to use a target resolution of 10 m or lower.
         debug_logs (bool, optional): If True, prints progress and debug info.
-        image_parallel_workers (ImageParallelWorkers, optional): Enables parallel execution. Note: "process" does not work on macOS due to PyTorch MPS limitations.
+        image_threads (Literal["cpu"] | int | None): Enables parallel execution. Note: "process" does not work on macOS due to PyTorch MPS limitations.
         omnicloud_kwargs (dict | None): Additional arguments forwarded to predict_from_array.
 
     Raises:
@@ -52,7 +47,9 @@ def create_cloud_mask_with_omnicloudmask(
 
     print("Start omnicloudmask")
     Universal.validate(
-        input_images=input_images, output_images=output_images, debug_logs=debug_logs
+        input_images=input_images,
+        output_images=output_images,
+        debug_logs=debug_logs
     )
 
     input_image_paths = _resolve_paths(
@@ -66,13 +63,14 @@ def create_cloud_mask_with_omnicloudmask(
             "default_file_pattern": "$_CloudClip.tif",
         },
     )
-    image_parallel, image_backend, image_max_workers = _resolve_parallel_config(
-        image_parallel_workers
-    )
+
+    # Determine multiprocessing and worker count
+    image_backend = "thread" # "thread" or "process"
+    image_threads_on, image_thread_workers = _resolve_parallel_config(image_threads)
+
 
     if debug_logs:
         print(f"Input images: {input_image_paths}")
-    if debug_logs:
         print(f"Output images: {output_image_paths}")
 
     image_args = [
@@ -89,8 +87,8 @@ def create_cloud_mask_with_omnicloudmask(
         for input_path, output_path in zip(input_image_paths, output_image_paths)
     ]
 
-    if image_parallel:
-        with _get_executor(image_backend, image_max_workers) as executor:
+    if image_threads_on:
+        with _get_executor(image_backend, image_thread_workers) as executor:
             futures = [
[…]
