### 2025-11-15 95dec1d
Replace Rasterio with GDAL for raster handling
Refactored code to replace Rasterio with GDAL for raster operations, simplifying dependencies and leveraging GDAL's advanced features. Updated functions for resolution, nodata value handling, and parallel processing. Removed deprecated utilities and added thread configuration options.
