+++ spectralmatch/types_and_validation.py
@@ -10,13 +10,10 @@ class Universal:
     SaveAsCog = bool  # Default: True
     DebugLogs = bool  # Default: False
     VectorMask = Tuple[Literal["include", "exclude"], str, Optional[str]] | None
-    WindowSize = int | Tuple[int, int] | Literal["internal"] | None
-    WindowSizeWithBlock = int | Tuple[int, int] | Literal["internal", "block"] | None
+    WindowSize = int | None
     CustomNodataValue = float | int | None
-    ImageParallelWorkers = (
-        Tuple[Literal["process", "thread"], Literal["cpu"] | int] | None
-    )
-    WindowParallelWorkers = Tuple[Literal["process"], Literal["cpu"] | int] | None
+    Threads = Literal["cpu"] | int | None
+    Cache = float | None
     CalculationDtype = str
     CustomOutputDtype = str | None
     CreateNameAttribute: Tuple[str, str] | None
@@ -31,12 +28,15 @@ class Universal:
         vector_mask=_UNSET,
         window_size=_UNSET,
         custom_nodata_value=_UNSET,
-        image_parallel_workers=_UNSET,
-        window_parallel_workers=_UNSET,
         calculation_dtype=_UNSET,
         custom_output_dtype=_UNSET,
         create_name_attribute=_UNSET,
         output_dtype=_UNSET,
+        cache=_UNSET,
+        image_threads=_UNSET,
+        io_threads=_UNSET,
+        tile_threads=_UNSET,
+        estimate_stats=_UNSET,
     ):
         if input_images is not _UNSET:
             if not isinstance(input_images, (str, list)):
@@ -65,20 +65,8 @@ class Universal:
             if save_as_cog:
                 if window_size is _UNSET or window_size is None:
                     raise ValueError("When save_as_cog=True, window_size must be set.")
-                if isinstance(window_size, int):
-                    if window_size % 16 != 0:
-                        raise ValueError(
-                            "When save_as_cog=True, window_size must be a multiple of 16."
-                        )
-                elif isinstance(window_size, tuple):
-                    if len(window_size) != 2 or window_size[0] != window_size[1]:
-                        raise ValueError(
-                            "When save_as_cog=True, window_size must be square (width == height)."
-                        )
-                    if any(w % 16 != 0 for w in window_size):
-                        raise ValueError(
-                            "When save_as_cog=True, window_size dimensions must be multiples of 16."
-                        )
+                if window_size % 16 != 0:
+                    raise ValueError("When save_as_cog=True, window_size must be a multiple of 16.")
 
         if debug_logs is not _UNSET:
             if not isinstance(debug_logs, bool):
@@ -101,59 +89,19 @@ class Universal:
                 )
 
         if window_size is not _UNSET:
-
-            def _validate_window_param(val):
-                if val is None or isinstance(val, int):
-                    return
-                if (
-                    isinstance(val, tuple)
-                    and len(val) == 2
-                    and all(isinstance(i, int) for i in val)
-                ):
-                    return
-                if val == "internal":
-                    return
-                raise ValueError(
-                    "window_size must be an int, (w, h) tuple, 'internal', or None."
-                )
-
-            _validate_window_param(window_size)
+            if window_size is None:
+                pass
+            elif isinstance(window_size, int):
+                if window_size <= 0:
+                    raise ValueError("window_size must be > 0.")
+            else:
+                raise ValueError("window_size must be an int or None.")
 
         if custom_nodata_value is not _UNSET:
             if custom_nodata_value is not None and not isinstance(
                 custom_nodata_value, (int, float)
             ):
                 raise ValueError("custom_nodata_value must be a number or None.")
-
-        if image_parallel_workers is not _UNSET:
[…]
