+++ spectralmatch/handlers.py
@@ -1,54 +1,13 @@
 import os
 import warnings
-import rasterio
 import re
 import glob
 
-from typing import List, Optional, Literal, Tuple, Union
+from osgeo import gdal
+gdal.UseExceptions()
+from typing import List, Optional, Literal, Tuple
 
 from spectralmatch.types_and_validation import Universal
-
-
-def _resolve_output_dtype(
-    dataset: rasterio.io.DatasetReader,
-    custom_output_dtype: Universal.CustomOutputDtype,
-):
-    """
-    Resolves the output dtype for a raster operation.
-
-    Args:
-        dataset (rasterio.io.DatasetReader): The input dataset to derive default dtype from.
-        custom_output_dtype (str | None): A user-specified output dtype, or None to use dataset dtype.
-
-    Returns:
-        str: The resolved output dtype.
-    """
-    if custom_output_dtype is not None:
-        return custom_output_dtype
-    return dataset.dtypes[0]
-
-
-def _resolve_nodata_value(
-    dataset: rasterio.io.DatasetReader, custom_nodata_value: Universal.CustomNodataValue
-) -> float | int | None:
-    """
-    Determine the appropriate nodata value for a raster dataset.
-
-    Priority is given to a user-provided custom nodata value. If not provided, the function attempts to use the nodata value defined in the dataset metadata. Returns None if neither is available.
-
-    Args:
-        dataset (rasterio.io.DatasetReader): The opened raster dataset.
-        custom_nodata_value (float | int | None): Optional user-defined nodata value.
-
-    Returns:
-        float | int | None: The resolved nodata value, or None if unavailable.
-    """
-    if custom_nodata_value is not None:
-        return custom_nodata_value
-    elif dataset.nodata is not None:
-        return dataset.nodata
-    else:
-        return None
 
 
 def _resolve_paths(
@@ -70,7 +29,9 @@ def _resolve_paths(
     """
     kwargs = kwargs or {}
 
-    if isinstance(input, list):
+    if mode == "name":
+        resolved = [os.path.splitext(os.path.basename(p))[0] for p in input]
+    elif isinstance(input, list):
         resolved = input
     elif mode == "search":
         resolved = search_paths(input, **kwargs)
@@ -78,8 +39,6 @@ def _resolve_paths(
         resolved = create_paths(input, **kwargs)
     elif mode == "match":
         resolved = match_paths(**kwargs)
-    elif mode == "name":
-        resolved = [os.path.splitext(os.path.basename(p))[0] for p in input]
     else:
         raise ValueError(f"Invalid mode: {mode}")
 
@@ -271,89 +230,67 @@ def _check_raster_requirements(
     if debug_logs:
         print(f"Found {len(input_image_paths)} images")
 
-    datasets = [rasterio.open(p) for p in input_image_paths]
+    datasets = [gdal.Open(p, gdal.GA_ReadOnly) for p in input_image_paths]
 
-    ref_crs = datasets[0].crs
-    ref_count = datasets[0].count
-    ref_res = datasets[0].res
-    ref_nodata = (
-        [datasets[0].nodata] * ref_count
-        if datasets[0].nodata is not None
-        else [None] * ref_count
-    )
+    ref_ds = datasets[0]
+    ref_crs = ref_ds.GetProjectionRef()
+    ref_count = ref_ds.RasterCount
+    ref_gt = ref_ds.GetGeoTransform()
+    ref_res = (ref_gt[1], abs(ref_gt[5]))
+    ref_nodata = [ref_ds.GetRasterBand(b + 1).GetNoDataValue() for b in range(ref_count)]
 
     for i, ds in enumerate(datasets):
-        if check_geotransform and ds.transform is None:
[…]

[…]
