### 2025-11-15 3dc7a69
Refactor and enhance benchmarking scripts and examples
Refactored raster creation and benchmarking logic, replacing rasterio with GDAL for better performance and streamlined overlapping raster generation. Updated threading parameters and parallel processing options in examples for consistency and improved configurability. Added new options like Cloud Optimized GeoTIFF support and enhanced debug logging.
