+++ docs/examples/benchmark.py
@@ -1,124 +1,113 @@
-import os, shutil, tempfile, time
-from pathlib import Path
-
-import matplotlib.pyplot as plt
+import os, time, tempfile, glob
 import numpy as np
-import rasterio
-from rasterio.transform import from_origin
-
+import matplotlib.pyplot as plt
+from osgeo import gdal, osr
 from spectralmatch import global_regression, local_block_adjustment
 
 
-def make_fake_rasters(out_dir, n_images, width, height, nodata=0):
-    out_dir = Path(out_dir)
-    out_dir.mkdir(parents=True, exist_ok=True)
-    profile = dict(
-        driver="GTiff",
-        width=width,
-        height=height,
-        count=8,
-        dtype="uint16",
-        nodata=nodata,
-        crs="EPSG:3857",
-        transform=from_origin(0, 0, 1, 1),
-        tiled=True,
-        blockxsize=512,
-        blockysize=512,
-        compress="LZW",
-    )
-    rng = np.random.default_rng(seed=42)
-    paths = []
-    for i in range(n_images):
-        p = out_dir / f"fake_{i+1}_{width}px.tif"
-        with rasterio.open(p, "w", **profile) as dst:
-            for b in range(1, 9):
-                data = rng.integers(1, 1000, size=(height, width), dtype="uint16")
-                data[0, 0] = nodata
-                dst.write(data, indexes=b)
-        paths.append(str(p))
-    return paths
+def make_two_rasters(size, out_dir):
+    """Create two square rasters that overlap by 50% in X."""
+    def _create(path, x0):
+        drv = gdal.GetDriverByName("GTiff")
+        ds = drv.Create(path, size, size, NUM_BANDS, gdal.GDT_UInt16,
+                        options=[
+                            "TILED=YES",
+                            f"BLOCKXSIZE={BLOCK}", f"BLOCKYSIZE={BLOCK}",
+                            "COMPRESS=LZW", "BIGTIFF=YES"
+                        ])
+        srs = osr.SpatialReference(); srs.ImportFromEPSG(EPSG)
+        ds.SetProjection(srs.ExportToWkt())
+        ds.SetGeoTransform((x0, PIX, 0, 0, 0, -PIX))
+        rng = np.random.default_rng(42)
+        for b in range(1, NUM_BANDS+1):
+            ds.GetRasterBand(b).WriteArray(rng.integers(1, 1000, size=(size, size), dtype=np.uint16))
+        ds.FlushCache(); ds = None
 
+    a = os.path.join(out_dir, f"A_{size}.tif")
+    b = os.path.join(out_dir, f"B_{size}.tif")
+    make_input = lambda p, xoff: _create(p, xoff)
+    make_input(a, 0.0)
+    make_input(b, (size * PIX) / 2.0)  # 50% overlap shift
+    return [a, b]
 
-SIZES = [2_048, 4_096, 6_144, 8_192, 10_240, 12_288]
-NUM_IMAGES = 2
-TILE_SIZE = (1024, 1024)
-MAX_WORKERS = 32
+def run_one(size, base):
+    """Create inputs, run global+local, return (t_global, t_local)."""
+    inp = os.path.join(base, "Input"); os.makedirs(inp, exist_ok=True)
+    gdir = os.path.join(base, "GlobalMatch"); os.makedirs(gdir, exist_ok=True)
+    ldir = os.path.join(base, "LocalMatch"); os.makedirs(ldir, exist_ok=True)
 
-WORK_DIR = Path(__file__).parent / "bench_output"
-WORK_DIR.mkdir(exist_ok=True)
[…]
