### 2025-11-15 9071015
Refactor tests for parametrization and remove legacy functions
Streamlined tests by introducing parameterized test cases for `merge_rasters` and `mask_rasters`, reducing redundancy. Removed outdated test functions and CLI commands for `create_ndvi_raster` and `threshold_raster`, aligning with updated functionality. Simplified function calls with unified argument names for clarity and consistency.
