+++ tests/test_match.py
@@ -30,9 +30,10 @@ def test_global_regression_full_options_save_model(tmp_path):
         calculation_dtype="float64",
         output_dtype="uint16",
         custom_nodata_value=0,
-        image_parallel_workers=("process", 2),
-        window_parallel_workers=("process", 1),
-        window_size=(16, 16),
+        io_threads=2,
+        image_threads=2,
+        tile_threads=2,
+        window_size=16,
         save_as_cog=True,
         specify_model_images=("include", ["A"]),
         custom_mean_factor=1.0,
@@ -74,8 +75,9 @@ def test_global_regression_full_options_load_model(tmp_path):
         output_images=new_output_paths,
         calculation_dtype="float32",
         load_adjustments=str(model_path),
-        image_parallel_workers=("thread", 1),
-        window_parallel_workers=None,
+        io_threads=2,
+        image_threads=2,
+        tile_threads=2,
         window_size=None,
         debug_logs=True,
     )
@@ -109,9 +111,10 @@ def test_local_block_adjustment_all_params_save(tmp_path):
         calculation_dtype="float64",
         output_dtype="uint16",
         custom_nodata_value=99,
-        image_parallel_workers=("thread", 2),
-        window_parallel_workers=("process", 1),
-        window_size=(16, 16),
+        io_threads=2,
+        image_threads=2,
+        tile_threads=2,
+        window_size=16,
         save_as_cog=True,
         number_of_blocks=(2, 2),
         alpha=0.75,
@@ -160,8 +163,9 @@ def test_local_block_adjustment_all_params_load(tmp_path):
         calculation_dtype="float32",
         output_dtype=None,
         load_block_maps=(str(ref_map), [str(p) for p in local_maps]),
-        image_parallel_workers=None,
-        window_parallel_workers=None,
+        io_threads=2,
+        image_threads=2,
+        tile_threads=2,
         debug_logs=True,
     )
 
