### 2025-11-15 56fd7ed
Update gdal dependency and update project URLs
Bumped gdal version requirement to 3.11 for compatibility reasons. Adjusted project URLs to reflect the updated repository location and issue tracker.
