### 2025-11-15 21ecdce
Remove redundant raster size validation in global regression.
The check for raster size equality was unnecessary and has been removed for simplicity. Updated the associated logic to ensure consistency in size usage across computations.
