### 2025-11-15 2305973
Remove unused seamline parameters and update raster dimensions.
Removed `min_point_spacing` and `min_cut_length` parameters from `test_seamline` as they're no longer needed. Adjusted raster dimensions in `test_mask` to improve test coverage and consistency, changing width and height from 5 to 64.
