### 2025-11-15 6adefb1
Refactor conditional check for debug_vectors_path
Revised the conditional statement to ensure `debug_vectors_path` is properly validated before attempting file operations. This improves code clarity and prevents potential errors caused by invalid paths.
