### 2025-11-18 spectralmatch/spectralmatch#49 @simonrp84
Hi,
My understanding is that because I am creating a fresh environment there should be no other packages installed. I can double check!
