### 2025-11-25 #8974 @cankanoa
Aloha, just following up here about this issue: https://github.com/spectralmatch/spectralmatch/issues/49. It has been resolved which means I believe we are just waiting on @simonrp84's review. Thanks to everyone for helping thus far.
