### 2025-12-30 e0e4afc
Add "offset" correction method for local radiometric adjustment
Introduced a new "offset" correction method alongside "gamma" and "linear" for local block-based adjustments. This method addresses images with negative values by applying a calculated offset to ensure valid logarithmic operations. Updated related logic, parameter handling, and documentation accordingly.
