+++ spectralmatch/match/local_block_adjustment.py
@@ -32,7 +32,7 @@ def local_block_adjustment(
     save_as_cog: Universal.SaveAsCog = False,
     number_of_blocks: int | Tuple[int, int] | Literal["coefficient_of_variation"] = 100,
     alpha: float = 1.0,
-    correction_method: Literal["gamma", "linear"] = "linear",
+    correction_method: Literal["gamma", "linear", "offset"] = "linear",
     save_block_maps: Tuple[str, str] | None = None,
     load_block_maps: (
         Tuple[str, List[str]] | Tuple[str, None] | Tuple[None, List[str]] | None
@@ -59,7 +59,7 @@ def local_block_adjustment(
         save_as_cog (bool): If True, saves output as a Cloud-Optimized GeoTIFF using proper band and block order.
         number_of_blocks (int | tuple | Literal["coefficient_of_variation"]): int as a target of blocks per image, tuple to set manually set total blocks width and height, coefficient_of_variation to find the number of blocks based on this metric.
         alpha (float, optional): Blending factor between reference and local means. Defaults to 1.0.
-        correction_method (Literal["gamma", "linear"], optional): Local correction method. Defaults to "gamma".
+        correction_method (Literal["gamma", "linear", "offset"], optional): Local correction method. Defaults to "gamma". Offset is commended for images with negative values.
         save_block_maps (tuple(str, str) | None): If enabled, saves block maps for review, to resume processing later, or to add additional images to the reference map.
             - First str is the path to save the global block map.
             - Second str is the path to save the local block maps, which must include "$" which will be replaced my the image name (because there are multiple local maps).
@@ -559,13 +559,14 @@ def _apply_adjustment_process_image(
     num_col: int,
     nodata_val: float,
     alpha: float,
-    correction_method: Literal["gamma", "linear"],
+    correction_method: Literal["gamma", "linear", "offset"],
     calculation_dtype: str,
     output_dtype,
     debug_logs: bool,
     tile_thread_on: bool,
     tile_thread_workers: int,
     save_as_cog: bool,
+    estimate_stats: bool = False,
 ):
     """
     Apply local radiometric adjustment (linear or gamma) to a raster image using block-based reference and local mean surfaces. Builds parameter surfaces as rasters, warps them to the image grid, and creates a VRT with per-pixel expressions, then materializes the output as GTiff or COG.
@@ -641,6 +642,23 @@ def _apply_adjustment_process_image(
         if ds_out is None:
             raise RuntimeError("Warp failed building parameter surface")
 
+    # Build offset if enabled
+    if correction_method == "offset":
+        global_min = np.inf
+        src = gdal.Open(img_path, gdal.GA_ReadOnly)
+        for b in range(1, num_bands + 1):
+            band = src.GetRasterBand(b)
+            min_val, _, _, _ = band.GetStatistics(1 if estimate_stats else 0, 1)
+            global_min = min(global_min, min_val)
+        src = None
+        global_min = min(
+            global_min,
+            np.nanmin(block_reference_mean),
+            np.nanmin(block_local_mean),
+        )
+        offset = abs(global_min) + 1.0 if global_min <= 0 else 0.0
+        offsets = [offset] * num_bands
+
[…]
