### 2026-01-06 #8974 @ethanwhite
@cankanoa - thanks to your excellent work based on the previous reviewer feedback things are looking really good. I opened a couple of minor PRs and one issue:

- https://github.com/spectralmatch/spectralmatch/pull/55
- https://github.com/spectralmatch/spectralmatch/pull/57
- https://github.com/spectralmatch/spectralmatch/issues/56

Based on some issues I ran into during installation and a missing set of imports in the README example. Take a look at these and see what you think. Once you've addressed them I can take off my reviewer hat and move on to editorial checks.
