### 2026-01-07 38f7dbd
Update spectral profile comparison to use GDAL for stats
Replaced rasterio with GDAL to calculate band statistics for spectral profiles. Added an optional `estimate_stats` parameter to control whether statistics are estimated. This improves compatibility and efficiency for handling image data.
