### 2026-01-07 28a9fb0
Add `rasterio` and reorganize dependencies.
`rasterio` was added to the Conda environment in workflow files and categorized under optional system-level dependencies in `pyproject.toml`. This improves clarity and ensures consistent environment setup across systems. Other minor adjustments enhance dependency management.
