### 2026-01-07 #8974 @cankanoa
Aloha @ethanwhite, Thanks for putting on the review hat to make this go forward. I've addressed all three points above: both PRs have been accepted and I added Ipython as a dependency. Let me know if there are any other ways I can help move this forward. Thanks for working with me on this so long.
