+++ docs/examples/example_worldview_mosaic.py
@@ -110,6 +110,7 @@ merge_rasters(
     input_images=clipped_folder,
     output_image_path=os.path.join(working_directory, "MergedImage.tif"),
     debug_logs=debug_mode,
+    window_size=window_size,
     io_threads=io_threads,
     tile_threads=tile_threads,
 )
