### 2026-01-12 3b150af
Add `window_size` parameter to mosaic functions
This commit introduces the `window_size` parameter to improve tile-based processing for raster mosaicking. It enables better control over tile sizes during raster creation and enhances the flexibility of the mosaic workflow. Minor adjustments to the function documentation and GDAL operation configurations are also included.
