+++ spectralmatch/utils.py
@@ -325,6 +325,7 @@ def merge_rasters(
     output_dtype: Universal.CustomOutputDtype = None,
     custom_nodata_value: Universal.CustomNodataValue = None,
     resolution: Literal["highest", "average", "lowest"] = "highest",
+    window_size: Universal.WindowSize = None,
 ) -> str:
     """
     Merges multiple rasters into a single output using GDAL Warp (C++), aligning them to the union extent and a unified resolution. Supports parallelism via GDAL threading knobs.
@@ -339,14 +340,11 @@ def merge_rasters(
         output_dtype (str | None, optional): Data type for output rasters. Defaults to input image dtype.
         custom_nodata_value (float | int | None, optional): Overrides detected NoData value. Defaults to None.
         resolution ("highest" | "average" | "lowest", optional): Strategy for computing merge resolution.
-
+        window_size (int | None): Tile size for processing tiles. Defaults to None.
     Returns:
         str: Path of the merged raster.
 
     """
-
-    if debug_logs:
-        print("Start raster merging")
 
     Universal.validate(
         input_images=input_images,
@@ -355,81 +353,63 @@ def merge_rasters(
         io_threads=io_threads,
         tile_threads=tile_threads,
         output_dtype=output_dtype,
+        window_size=window_size,
         custom_nodata_value=custom_nodata_value,
     )
+
+    # Setup parallel
+    tile_thread_on, tile_thread_workers = _resolve_parallel_config(tile_threads)
 
     input_image_paths = _resolve_paths(
         "search", input_images, kwargs={"default_file_pattern": "*.tif"}
     )
 
-    output_dtype = _resolve_gdal_dtype(output_dtype, input_image_paths[0], debug_logs)
-    nodata_val = _resolve_nodata_value(input_image_paths[0], custom_nodata_value)
-
-
-    # Setup parallel
-    tile_thread_on, tile_thread_workers = _resolve_parallel_config(tile_threads)
-
+    # Dtype
+    output_dtype = _gdal_dtype_str_to_enum(_resolve_gdal_dtype(output_dtype, input_image_paths[0]))
 
     _set_gdal_cache(cache, debug_logs)
     _set_gdal_workers(io_threads, debug_logs)
 
     if debug_logs:
-        print(f"Merging {len(input_image_paths)} rasters to: {output_image_path}")
+        print(f"Building VRT from {len(input_image_paths)} rasters")
 
-    # Compute resolution using strategy
-    xres, yres = compute_resolution(input_image_paths, resolution)
+    vrt_opts = gdal.BuildVRTOptions(
+        resolution=resolution,
+        srcNodata=custom_nodata_value,
+        VRTNodata=custom_nodata_value,
+    )
 
-    # Compute union bounds
-    bounds_list = []
-    for path in input_image_paths:
-        ds = gdal.Open(path, gdal.GA_ReadOnly)
-        gt = ds.GetGeoTransform()
-        w, h = ds.RasterXSize, ds.RasterYSize
-        xmin = gt[0]
-        ymax = gt[3]
-        xmax = xmin + w * gt[1]
-        ymin = ymax + h * gt[5]
-        bounds_list.append((xmin, ymin, xmax, ymax))
-        ds = None
+    vrt_ds = gdal.BuildVRT("", input_image_paths, options=vrt_opts)
 
-    xmin = min(b[0] for b in bounds_list)
-    ymin = min(b[1] for b in bounds_list)
-    xmax = max(b[2] for b in bounds_list)
-    ymax = max(b[3] for b in bounds_list)
+    creation_options = [
+        "TILED=YES",
+        "BIGTIFF=YES",
+        "COMPRESS=ZSTD",
[…]
