### 2026-01-12 8db7088
Update legend customization in plotting function
Refined legend placement and appearance for better layout and readability. Added `tight_layout()` call to prevent overlapping elements and ensure proper spacing in the plot.
