### 2026-01-17 #8974 @cankanoa
Yes that makes sense, I'm having issues with submitting a Zenodo archive. For some reason uploading the files is stuck on 'pending', I've started an issue with them. Hopefully this is resolved quickly. I will keep you updated.
