### 2026-01-21 f3da11a
Add matplotlib as a required dependency
This update includes matplotlib (version 3.8 or higher) in the required dependencies section.
