+++ pyproject.toml
@@ -38,6 +38,7 @@ dependencies = [
     "scikit-image>=0.25.2",
     "opencv-python>=4.11.0.86",
     "fire>=0.7.0",
+    "matplotlib>=3.8",
 ]
 
 [project.optional-dependencies]
