### 2026-01-21 #8974 @cankanoa
@ethanwhite I've created the release and archive. Here is the [archive](https://zenodo.org/records/18330320) and here is the [release](https://github.com/spectralmatch/spectralmatch/releases/tag/v1.0.4). Let me know if there is anything else I can do to finish things up. Thanks so much for working with me on this.
