### 2026-01-22 #8974 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
