### 2026-01-23 #8974 @kthyng
Congratulations on your new publication @cankanoa! Many thanks to editor @ethanwhite and to reviewers @Digdgeo and @ethanwhite for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts. Thanks to @ethanwhite for an extra time investment on this submission as both editor and reviewer!
